Start the tree with predicates
This initial commit implements every box in
the decision tree as a predicate, and provides
elementary unit testing. There is no execution
flow through the decision tree in this commit.

Boxes (decisions) contingent on statute, legal
or tax code are left implicit for now. The data
informing the decision is defined solely via
the appropriate specs.