Refactor allowed FERPA exceptions
This change is Cursor implemented with some fairly
granular prompting to direct the agent appropriately.

Rubocop also suggested some refactoring via an
autocorrent rule, which was applied. This change
is an optimization in the logic: the financial
aid information can always be disclosed to the
student regardless of the presence of FTI. The
diagram could be simplified a little bit using
this fact.